Remove call to WikibaseRepo::getUserLanguage from service wiring
In support of the deprecation and eventual removal of
WikibaseRepo::getUserLanguage, remove calls to the service from the
MessageParameterFormatter service wiring in Wikibase.

The existing MessageParamaterFormatter service takes a language
object as a parameter. Since our ultimate goal (T279415) is to drop
such request-specific dependencies from services, we rewrite the
MessageParameterFormatter here to emit language-independent wikitext
instead of localized text.

For formatting numbers, we introduce a new parameter to the core
`formatnum` parser function which implements the accuracy-
preserving behaviour of the `MediaWikiNumberLocalizer`. (T393483)

For formatting lists, we can use Wikibase's `#commaSeparatedList`
function.

Bug: T281720
Depends-On: I2372dcec6f890f60efe42acf30db0640dc685117
Change-Id: I650c87110b5085a9df5c1967f8f4e408ec281c05