repo: Add generate= parameter to wbformatentities API
The parameter is meant to resemble that of wbformatvalue, with the same
name and (subset of) values, though behind the curtain it actually works
somewhat differently – we’re not passing the format to a generic
formatter factory that supports multiple output formats, there are just
two different formatters for the two supported formats. If anyone wants
a FORMAT_WIKI version, or even FORMAT_HTML_VERBOSE(_PREVIEW), we’d need
some more work (with a new formatter) to support that, but for now I
don’t think it’s needed.

Bug: T393691
Change-Id: If69c4276dcc70df4a09ed01164a38c232835145a