Add new mobile statement UI config flag
This adds a new config setting for the mobile editing UI. The flag is
passed through into the ItemView, where we will use it to toggle between
the “regular” statements view and the new version (just a placeholder
for now), and also adds a new ResourceLoader module (temporarily
registered in RepoHooks, like in change I7d8b53f9a8, commit df15eb3595).
For now, the new mobile UI is only controlled by this flag – the beta
feature added in change Ib76c441caf (commit c8ead6876d) is not taken
into account yet.

FullEntityParserOutputGenerator gains a second call to record the
wbMobile parser option (compare the existing call in addModules()),
because now not just the ResourceLoader modules but also the HTML itself
depends on the mobile flag. For now, we don’t add the mobile UI flag
itself to the parser options, though that might happen later.

The config flag itself is not recorded in the parser options (that might
also happen later), so visitors may still see the real statement UI on
mobile after this feature flag has been turned on, or the fake UI after
it has been turned off, based on old parser cache entries. However, note
that on dev wikis this is likely to be affected by other factors:

- Parser output with broken Termbox SSR is not cacheable, so to have a
  mobile parser cache for items at all you’d need to either set up
  working SSR or set $wgWBRepoSettings['termboxEnabled'] = false;
- $wgInvalidateCacheOnLocalSettingsChange is true by default, so
  MediaWiki will effectively reset the parser cache each time you touch
  LocalSettings to toggle the feature flag

Bug: T394291
Change-Id: I3b61c006ecef8c37ca9890c7adccdd6ee29929cd