Remove WikibaseClient.UserLanguage service
It’s only used in one place, so inline it there. I think the benefit of
removing the UserLanguage service (and thus preventing anyone else from
adding more usages of it) outweighs the badness of the other service
wiring now using $wgLang.

This doesn’t resolve T344050, it’s just related to it. Eventually we
should still find a way to fix the badge display (but I don’t know how).

Bug: T344050
Change-Id: I412face9c6e8fd81eef4c6c988ca3937e4b20c53