Add a static example of a VueJS mobile statement view
For the MEX mobile editing changes, add an example of
a simple static statement view. This is not yet
populated with data (T395791), and tests will be added
in a follow-up change (T395795). Subsequent changes will
also take care of presenting a useful no-JS / server-side-
rendered view of the data (T395802).

Bug: T394885
Change-Id: I2a47dea3ae185243f1e27018f716b7d620d52e6a