Fix edge-cases in SummaryParsingPrefetchHelper
I found a bunch of potential issues in this regex.

* It does not only find [[P1]] but also [[PPP1]]. That will be
rejected by the entity id parser just a moment later and doesn't
cause problems. But it's wasting time and resources doing that.

* Link labels like in [[…|P1]] are picked up as if they are titles,
but they are not.

* The regex was phrased in a way that only whitespace would work
as seperator between links. In something like [[P1]]-[[P2]] neither
of the two links would be found.

The modified test cases cover this now.

Bug: T292545
Bug: T388685
Change-Id: Ib39b8910644738b10dd8f43c30a981df6976770a