The AbortEmailNotificationHook gets deprecated, use Middleware instead
MediaWiki core finished migrating Watchlist ENotif feature into a
generic notification platform. As a part of this work multiple hooks
related to ENotif are deprecated in favour of a Middleware pattern.

To abort the Notification from happening we need to inject a
FilterMiddleware that allows to suppress notification. But if required
Wikibase can not only suppress but also modify/update the recipients
list.

Bug: T395352
Change-Id: I3efee7942b6916cb721bb19a8cb4cdf9879f5753