Use php-vuejs-templating to render new mobile UI
This library (which we already use in WikibaseLexeme) can parse the
template and perform a subset of Vue server-side rendering, such as
basic v-bind: and {{ text }} substitutions. We will try to use that to
provide a read-only no-JS experience for the new mobile UI.

To make the icons work, turn them into CSS-only versions [1], and use
the CSS-only version in JavaScript too. (We could add support for
handling <custom-elements> to php-vuejs-templating, and we might still
end up doing that, but it seems pointless in this case. If we write
<cdx-icon> in the template and teach php-vuejs-templating to emit it as
a <span> with a suitable class, it just means that we have to send the
actual icon over the wire twice – once in CSS and once in JS.)

[1]: https://doc.wikimedia.org/codex/latest/components/demos/icon.html#css-only-version

Bug: T395802
Change-Id: I68a627b3c831c8c28392beb0b944c4b06424a49c
Depends-On: I6a5562644ab90acaa8e1069de840ac2ddd256502
Depends-On: I77a8ef50237cb051e1ad520a7a871550f740ad9e
Depends-On: I9d43a495fbe3e636be1bf62fbcac93a2e39d7bb1