Bring real data into the mobile statement template
This commit modifies the server-side rendering to include some of
the information about the item's statements, and provides statement
elements that the vue components can mount upon.

The vue component now accepts a `statement` property. One component
per property group is mounted on the server-side-rendered elements,
and uses content from the first statement in the template.

This commit also includes the first jest tests, which validate that
the component mounts correctly.

Bug: T395791
Bug: T395795
Change-Id: I054a2bc301c46cc56d5caf6526325f5ee433ecf0