Remove @nomin pragma
We are removing the use of nomin because we feel it's no longer needed after the rollout of ES2017 to mediawiki.

Why was this added initially?
- The use of @nomin was a way to avoid minification of tokens we didn't yet support in MediaWiki. Parsing tokens we didn't support could result in invalid JavaScript output after minification.

The files affected in this patch have been verified with ESLint and they pass ES2017 which we now support. It might be useful to note that data-bridge.app.js didn't pass ES2015 because of the use of async. But that doesn't matter since it's supported in ES2017.

Bug: T378498
Change-Id: I111d0aeb9d37aac36bdbd2e946fcbfb09877b991