Allow reading local short description via Lua mw.title objects
Lua modules can use the shortDescription attribute on an mw.title object
to query the local short description for the page. This will increment
the expensive function counter, as well as register the page as a
transclusion so that the parser cache is refreshed on edits.

A parser test is added as well.

Bug: T216356
Change-Id: I11868336e9ed7728f4775f58639a77f41118d6ca