Improve property label and link in mobile statement template
This version should be compatible with php-vuejs-templating
2.2.0-beta.*. The cost is that we have to duplicate some of the logic
between PHP and JavaScript, but I think that had to happen anyway (even
if we extend php-vuejs-templating’s JavaScript support substantially,
it’s never going to support mw.Title()).

Some TODOs are still left; I don’t know what the right way to look up
the property namespace number is, but the code here probably won’t work
on non-English Wikibases. And the property label is of course still
completely missing. (We could look it up in the SSR version, but we
might as well leave that for later.)

Bug: T396633
Bug: T396634
Change-Id: I2f9a0a9165c8187f01f0d68cf73f61b242a004a5