Extract first subcomponent for mobile editing experience
Using the new App class in the upcoming php-vuejs-templating 2.2.0
(currently -beta.2, will need more polishing before a stable release).

Note that the props for the property name component currently must be
called url and label, not propertyUrl and propertyLabel. This is because
attribute names are lowercased during HTML parsing in PHP, so we must
specify them using kebab-case (:property-url=, not :propertyUrl=); but
php-vuejs-templating doesn’t currently undo the kebab case conversion
(so inside the subcomponent, {{ propertyUrl }} would be missing because
it only sees a property-url variable). A future php-vuejs-templating
version will fix this [1], at which point we could rename the attributes
if we wanted to.

[1]: https://github.com/wmde/php-vuejs-templating/pull/37

Bug: T396863
Change-Id: I2fae398f716a9a0cdef390f4e89e2d5fbf908e3d
Depends-On: I5f2fd83f73b78d8c91c7831af9bead4f7635eb5e