Clicking the search button goes to Special:Search
There was a bug such that clicking the search button would bring
you to main page rather than Special:Search with the results of that
search. The cause of this was that the `Search` button was not actually
part of the ScopedTypeaheadSearch component, and was instead a residual
button element from the basic HTML/CSS version of the page. As it had
no relation to the entity type selection or the search input field, it
could not do anything useful.

This commit sets the button-relevant props on the Codex TypeaheadSearch
component, and tidies up the Vue mounting process and css to keep the UI
the same as before.

Bug: T397506
Change-Id: Icd22b66c407a5bcf3268fe198539188d4fdf7a7a