Add initial draft of commonsMedia statement view
Based on the recent support introduced in php-vuejs-templating
2.2.0-beta.2, add an initial draft of support for commonsMedia-type
entities.

The draft renders hard-coded dummy data - there is not currently
enough model data available to faithfully display a commonsMedia
item. But the layout of the image and text approximately match the
figma design.

A 'mainSnak' Vue file has been introduced to handle the logic of
displaying different entity types. Because of a) the presently
limited support for more complex Javascript expressions in Vue
templates (T396855) and b) the lack of support for computed properties
in Vue templates (T397223), we have to pass some additional properties
to the `mex-main-snak` component that will later be removed.

This draft also "fixes" the display of non-string, non-media items
to display an error message.

Bug: T394907
Change-Id: I516400cdf26c9ddf6d685236f3ffe71c6e6c0e36