Use statement serializer to load item data into Vue templates
The MEX ItemVue currently uses manually- and incompletely-serialized
statement data for the server-side rendering of the Vue templates.

Use a StatementSerializer so that the data is complete.

Because of the nested data in existing templates and a bug (T398192)
in 2.2.0-beta.3 of php-vuejs-templating, this can generate a server
error rendering item-valued statements until the bug has been fixed
and php-vuejs-templating upgraded.

Bug: T396858
Change-Id: Id49eaad56cec6ec980c5b0f241d83dd342d16721