Use property label in MEX SSR
Use the WikibaseRepo.EntityIdHtmlLinkFormatterFactory service to format
the property label server-side. This does a lot of things that I expect
we’ll want sooner or later: it formats the property as a complete link
with the right URL, it gets the label, it applies language fallbacks,
and it even shows a language fallback indicator. However, it uses
different class names, and we’ll have to either live with those (and
adjust our CSS) or teach it to emit wbui2025 class names instead; for
now, I’ve used the :link pseudo-class to style the links inside.

Bug: T396633
Change-Id: If1948a336b9257b2370ffb3fdee3cf56ecd8df12