Set `wbMobile` parser option to `wbui2025` for new mobile interface
In preparation for deploying the wbui2025 / MEX feature to Wikidata
beta, adjust the `wbMobile` ParserOption value so that renders of
the new UI land in distinct cache entries from renders of the
existing UI.

We currently only enable the setting (`tmpMobileEditingUI`) globally,
so it suffices to make the cache key dependent on the global value
of that setting. If we later decide to enable the feature on a per-
user basis (T394704), we will need to make the value of the
ParserOption (and the renderer selection) dependent on the current
user's settings.

Bug: T399703
Change-Id: Iae60023cf1151f7918f127cb522dd2c3ff1dbdf0