Refactor wbui2025 implementation to target StatementSectionsView
The current wbui2025 (MEX) implementation targets the ItemView
class and is only available for Items. The wbui2025 design is
currently focussed on the statement list, and the design changes
are common between the Property and Item mobile view.

Move the implementation of wbui2025 to the StatementSectionsView
so that the code can be reused between Item and Property (and
ensure that the same view is not applied to Lexemes or other
entity types, since that is out of scope for the current
milestones).

Bug: T400085
Change-Id: I9c78aa1e2a516b4044099e60eb0987d915408bae