Align uses of snak / data value HTML
Previously, we were using a mixture of SnakFormatter output and
SnakHtmlGenerator output for formatted snaks, and if the same snak was
used in a main snak and a qualifier or reference, it could result in a
bad display and a console warning. Use SnakFormatter output everywhere.

The SnakFormatter output, unlike the SnakHtmlGenerator output, does not
contain the property of the snak. This requires significant changes to
the templates and styles, because we now have to add the property link
ourselves; this is done using the propertyName component, but that
introduces a new <p> element around the property label, so several
styles are added to remove the padding of that again. Maybe later we can
remove the <p> from the propertyName component (and instead add it only
to the main snak display?) and remove some of these styles again. The
result isn’t a pixel-perfect match for the previous outcome, but I
assume that’s okay because the previous version wasn’t a pixel-perfect
match for the design either – it’ll be further refined later.

To try to reduce confusion, also rename snakHtml() to snakValueHtml();
we don’t really have a good name in Wikibase for “the non-property part
of a snak” (“value” is arguably misleading because it doesn’t cover the
“somevalue”/“novalue” options), but I guess this will do for now.

Bug: T400123
Change-Id: Ic99ef9e27479acd54bda6942462ed78e856e4dd4