GQL: Request labels in batch
This reduces the number of DB queries to fetch labels. It combines the
labels to fetch of the main item with the ones to fetch of statement
properties. In theory, this could result in over-fetching if item labels
are requested in different languages than the statement property labels,
but in reality this is unlikely and probably not worth optimizing.

Bug: T400146
Change-Id: Ie92b5bfaad562649c284b6de41d4d533c4903f6a