CRUD: Make deleted property tests less flaky
These tests sometimes fail in CI on the assertion that the data type of
the deleted property is `null`. This can happen if the data type lookup
cache still holds the outdated value after the property is deleted.
There is also a chance that the property deletion itself failed without
us noticing. This patch deals with both of these issues by 1) asserting
that the property deletion succeeded in the setup code, and 2) disabling
the cache when requesting the statement using the deleted property.

We thought that we had previously solved this problem by arranging the
test setup code in a way that invalidates the cache before the deleted
property is checked in I605956d74e24d0947202728ee3851d448a84c70b, but
the tests suddenly started failing again, seemingly after merging
I0815488c3dedb9eb09309d74c7fd9e8a4a4e31d6.

Bug: T400265
Change-Id: I57a28055db3600d0598ab7ae5d15e2df3b5124a2