Disable vue/no-v-html rule in wbui2025 directory
The v-html directive is part of our expected workflow for reusing
server-rendered HTML (property links, snak values; see the upcoming ADR
of T399832). As such, this eslint rule is not currently useful for us.
If eslint adds the ignorePattern (or similar) option suggested in [1],
then we may later use it again (but note that this will require us to
upgrade to the latest version of eslint-plugin-vue first).

[1]: https://github.com/vuejs/eslint-plugin-vue/pull/2857

Change-Id: I8926d09a67927dc36b0bd5cfdb3e5c6a7256842e