GQL: Allow filtering by statement property
This adds an optional "properties" filter to the statements field, so
that we can either get all statements by not providing the argument
(`statements { ... }`) or specify which statements we want by providing
a list of properties (`statements(properties: ["P31", "P279"] { ... }`).
This seems like a useful (and cheap) feature in general and makes our
tests nicer.

Validating Property IDs seems not necessary in this case, because
they're guaranteed to be a list of strings by the schema, and even if
the user sends nonsense, it just won't have an effect (garbage in ->
garbage out).

Change-Id: I8fa62b7daf3716d09d3df5ec2cb99af621269ebe