Statements with Globe Coordinate datatype
The Kartographer extension is responsible for displaying maps, converting
a `<mapframe>` tag into a static map before/without js, and adding
interactivity in its client-side code. When Kartographer initializes
interactive maps, it modifies the DOM by nesting several `<a>` tags
inside of a container `<a>` tag. While this structure can be produced
in the DOM, it is not valid html and is modified by the browser's
html parser when inserted via `innerHTML=` or Vue's `v-html` directive.

Prior to this commit, wbui2025's init code imported Kartographer's
invalid html into the Pinia store for later use in the Vue template.
The invalid html was "corrected" by closing the top-level `<a>` tag
as soon as it encountered another one nested inside. Those subsequent
`<a>` tags containing the interactive map controls were moved outside
of the map container.

This commit imports the html into pinia earlier in the process when
the `wikibase.wbui2025.entityViewInit` module is loaded, getting the
un-modified (or at least less-modified) server-rendered, *valid* html.

Upon mounting a WikibaseWbui2025MainSnak Vue component with the type
`globe-coordinate`, Kartographer's `initMapFromElement` is called on
the element, introducing the map's interactivity. This is similiar to
what is done for map previews in desktop editing.

Bug: T394906
Change-Id: Ie606d9fbe09281240dbb0587971e5c0e0684ad1c