Use native str_contains instead of strstr
strstr is one of the oldest, strangest PHP functions. It returns a
substring but is universally not used for that, but more often used
as if it's a boolean function. It's not. The new str_contains is.

This makes the code much more readable (I believe) as well as a little
bit faster.

Change-Id: I564beef7d7fcdc8cd4115e48d8fd9b542aabc265