Add a first simple test of the statement view
Using the cypress integration, add an Item pageObject and test to see
the statement list container is displayed in the mobile view.

This change introduces the first `wbui2025` spec, where we set the
`viewport` to be 375 px wide per the initial first-pass requirements.
We also start using the `MwApi:CreateItem` task provided by
`cypress-wikibase-api`, which has a dependency on `api-testing` (which
is already included in Wikibase to run end-to-end tests).

The latest version of `api-testing` is 1.7.1 which, since Ib9a0fc9c,
includes upgraded versions of supertest and superagent. Unfortunately,
the latest versions of these libraries have an issue which prevents
supertest being required from browser contexts ([1], [2]). Doubly
unfortunately, failing to upgrade supertest and superagent means that
Wikibase developers see deprecation warnings during `npm install`. The
deprecation warnings also point to a security issue that has been fixed
in superagent [3]. While our test code is unlikely to be presented with
any unsanitised input, and certainly never running in production, this
is clearly undesirable.

In the interests of introducing some tests now and making progress with
wbui2025, the patch pins `api-testing` at 1.7.0. Depending on how long
it takes for upstream to respond to the issues, we might need to
consider other approaches for creating test fixtures on target wikis.

1. https://github.com/forwardemail/supertest/issues/867
2. https://github.com/forwardemail/superagent/issues/1840
3. https://github.com/forwardemail/superagent/pull/1800

Bug: T400471
Depends-On: I186f4b145d38f1df8735415c4f9fb63529be3cbe
Change-Id: I6a6a69642926cbe60af04c393d1a1eb7aa60f595