Fix Phan errors due to Status becoming generic
Status was made a generic class in MediaWiki core change I76dacc5809
(commit a965c6d260), which caused various Phan errors in Wikibase. Fix
them again.

For the most part, this makes the typing stronger and should, overall,
be a good thing for the future. In the maintenance script, I decided to
just replace the old code (Maintenance::error() can take a Status and
format it correctly), because I have no idea if the old code was correct
(if attemptSave() fails, is the resulting Status guaranteed to have a
string as the value?).

The typing of EditEntityStatus gets a bit awkward. Ideally, we’d like to
inform Phan that the class is still generic, but that the type parameter
must extend array{savedTempUser:?UserIdentity,context:IContextSource}
(possibly adding further elements to the array); however, I don’t think
that’s possible without [1]. In lieu of that, pretend it’s a “fully
generic” class (with the type parameter requirement just documented in
prose), fill in the type parameter in the subclasses (EditEntityStatus
etc.), and manually annotate the getValue() return type in the two
TempUserStatus methods by extracting it into a variable. (Note that the
class has subclasses in WikibaseLexeme and EntitySchema, so follow-up
changes are necessary there because otherwise Phan will complain about
the missing type parameter in those classes.)

Also apply a PHP 8.0 improvement in TempUserStatus while we’re at it.

[1]: https://github.com/phan/phan/issues/1666

Bug: T401468
Change-Id: I1d6b82e2d23ad99f2a98d542c7f568545c4d8910
Depends-On: I76dacc5809aeefe6f287a227e28e3a57ccf01d67