Use mediawiki-phan-config 0.17.0
LibUp is currently broken for Wikibase (T391763), so if we want the new
Phan version (which will be useful for T397781), we have to update it
manually. There are only two errors anyway.

In ChangeHandler, Phan complained that we were returning a value of type
(?Title)[], incompatible with the declared return Title[]. This is fixed
easily by changing castFromPageIdentity() to newFromPageIdentity() – the
only difference between them is that the former returns null if given
null, but we know that the input page records are going to be non-null.

In RdfBuilder, Phan complained about the potential infinite recursion
due to using the same arguments (i.e., no arguments). The error message
itself points out that Phan doesn’t analyze any changed state outside
the function parameters that might limit the recursion, as indeed is the
case here; I’m guessing we’re supposed to just suppress the warning in
this case, as some other extensions already do [1].

[1]: https://codesearch.wmcloud.org/search/?q=PhanPossiblyInfiniteRecursionSameParams&files=\.php%24

Change-Id: Iec173ff49c88a45ead328ddac07d27405c47deb4