Add wbui2025 property selector
Appearance-wise, this is kind of a Frankenstein’s monster of several
versions of the UI in Figma – it has the buttons in the heading of [1],
but the normal Codex lookup appearance of [2] rather than the more
“full-screen” menu of [1]. It remains to be determined which of these we
should actually have.

The JS code for operating the Lookup is very largely based on
ScopedTypeaheadSearch.vue, but with some improvements (we can use
async/await now!) and simplifications (no namespace selector here).

Because this component is not used during server-side rendering, we can
use a <style> block in the SFC file itself instead of a separate .less
file.

All the messages are reused from the desktop UI; we might want to
introduce one or two new ones (e.g. wikibase-statementgrouplistview-add
is in initial lowercase, which might not be great for a heading), but
we’ll see about that later.

The visibleItemLimit in the menuConfig is an arbitrary value; we should
revisit that later. Also, it would be nice to pass the url into the menu
items so users can open them in a new tab (check if they’ve selected the
right property or not), but that’s blocked on T342507: currently, if we
were to do it, users couldn’t tap a property to select it (it would
register as following the link instead).

[1]: https://www.figma.com/design/RtMDwH7sMF5IPRJvyyZPv4/Mobile-editing-experience?node-id=1003-12673&m=dev
[2]: https://www.figma.com/design/RtMDwH7sMF5IPRJvyyZPv4/Mobile-editing-experience?node-id=1904-26284&m=dev

Bug: T401402
Change-Id: I2f0309a6875257a5de0afe9978176247806a71be