Remove unused eslint suppression in statementSections.vue
We already disable this rule in .eslintrc.js for all wbui2025 files.
This was missed in change I8926d09a67 (commit 47afac06a6) because
statementSections.vue was added concurrently to that change (it’s
present in the rebased-on-master commit, but was not present in the
original change uploaded to Gerrit before it was rebased during merge).

Change-Id: I41484e75db78809b9ef5dc90a713ecb42d30a870