Create a add statement button
One button per statement section (note that pages will need to be purged
to make the new -add-wrapper show up in the server-rendered HTML).

We don’t have a design for this button yet, so for now it’s just a plain
button with the same margin-top as the statement groups.

Bug: T401401
Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Change-Id: I7c14326832921e297030a557d23f3da7902a4f1d