Make edit link for unsupported data types red
And change the cursor to the default. Note that it’ll still get an
underline on hover (though I guess that’s less of an issue on mobile).

Tested at the unit level, rather than in Cypress, because here we can
control which data types are supported and unsupported (through the
supportedDatatypes.json mock), whereas in Cypress at some point all
available data types should be supported.

Bug: T402448
Change-Id: I1099e86d2516c494526bdb96ee09aba4d85d9fe4