Allow removing terms in unsupported languages
This will make it easier to phase out support for language codes that,
for one reason or another, should no longer be allowed in labels,
descriptions, and aliases: with this change, we can remove the language
code from the set of supported languages immediately, and Wikibase will
still allow removing terms in these languages (but not adding or editing
them).

Note that this only allows removing terms in unsupported languages using
the wbeditentity API (Action API) or a full-entity PATCH (REST API).
APIs that operate on specific languages (wbsetlabel, wbsetdescription,
wbsetaliases; DELETE /labels/{language_code}) are not supported – in the
case of the Action API, the error looks like this:

> Unrecognized value for parameter "language": [language code]

I think this is an acceptable restriction and trying to lift it is not
worth the added effort.

The new test for this in EditEntityTest (adjusting an existing test case
rather than adding another one to this already slow test class) uses a
new config argument to change the $wgExtraLanguageNames just for one
test; this replaces the previous $needed argument which was already
unused since change I3d01f6ed87 (commit 73b4ac268c). (Also, if the test
fails for you locally because "en-x-T403097" is too long for the
wbxl_language column, you may need to apply the schema change from
I1c1f2b729f to your dev wiki, as I had to.)

There were some existing tests which checked that removing terms in
invalid language codes isn’t allowed; however, as far as I can tell
those were just added as part of invalid language code handling tests in
general (in change Ibb6ee8439b, commit ca42db59f4), so I don’t think
they represented an urgent product requirement at the time. Now that
we’re intentionally changing this behavior, I think it’s reasonable to
adjust the tests accordingly.

Bug: T403097
Change-Id: Ie034ae9ab89d1249510bfe65e523bc6203a06ed2