Rename wbui2025 editStatement to editStatementGroup
With the way the wbui2025 UI is designed, the “top-level unit” of an
edit is the statement group (we show a dialog for editing all the e.g.
“instance of” statements at once). Rename the component to match that,
as it contains more than one statement. (The individual statements are
represented by a component that’s currently named editStatementAddValue;
this is renamed accordingly in change I49bfbef09f.)

I’ve left editStatement.cy.ts alone in this change, because I think that
file name arguably doesn’t refer to the component called editStatement,
but rather to the action of editing a statement, and I think it’s fair
to say that this browser test is more about editing a single statement
than a statement group.

Bug: T403184
Change-Id: I6fe3fcc26a986b40255fedcfca63da58f91f9ff3