Rename wbui2025 editStatementAddValue to editStatement
This component contains not just the (main snak) value of a statement,
but also the qualifiers and references (and IIUC we’ll use it not just
for adding, but also for editing values). I think it deserves the
editStatement name, now that that’s been freed up by change I6fe3fcc26a.

Bug: T403184
Change-Id: I49bfbef09f55e44f947018cad3795e5d3fd3a804