Fix edit link appearance
Change I80ba8c7ba0 (commit 927aaf18fa) reorganized the link styles so
that the wikibase-wbui2025-link-heavy class no longer adds link styles
on its own, but only when used together with wikibase-wbui2025-link. (In
this regard, the nesting of the stylesheet is arguably misleading,
because the &-heavy class doesn’t actually get the “outer” styles at
all.) To make the edit links look like links again, add the missing
class to them, and also move the is-red-link class to the outer element
(because only that element has .cdx-mixin-link() now).

Bug: T400178
Change-Id: I1e40eaa30f590c9a1546fa63cb96d23846ace9ac