LanguageAwareRenderer: Don’t use Status
The Status never “escapes” this class; it’s simpler to just not use it
in the first place, and only deal with messages and wikitext.

While we’re here, also add more static types. (But not to render()’s
$propertyLabelOrId parameter, because that method is inherited from an
interface; we can’t narrow its parameter type without also doing it in
the interface, and I don’t feel like touching the interface right now.)

Bug: T401468
Change-Id: Id5f09c4dd16f45039ad3311749e87c9f0924a8ad
Co-Authored-By: Bartosz Dziewoński <dziewonski@fastmail.fm>
Co-Authored-By: Tacsipacsi <tacsipacsi@jnet.hu>