Respect $wgEditSubmitButtonLabelPublish
The “publish” message should only be used if this config is set to true;
by default, the “save” message should be used instead.

The JS variable is added in ParserOutputJsConfigBuilder, together with
wbEntityId, so we can rely on it being there and don’t need to send it
to the frontend separately.

In theory I guess we could tweak RepoHooks so that only the message
which will actually be used is sent to the frontend, but I don’t think
it’s worth it (both messages are short), especially considering we’ll
want to move the module definition to extension-repo.json eventually
(see T385446).

Bug: T402630
Change-Id: I10c06d713f6459bb142060fbeedb44601bd2e142