Refactor editStatementGroup Jest tests
The approach of mounting the component in a beforeEach() handler is
limited because it allows no test-specific customization before the
mounting – if we want to change anything, we have to copy the whole
mounting block (as done in e.g. statementGroupView.spec.js). Let’s try
an alternative approach where the mounting happens in a function.

Bug: T402630
Change-Id: I783783950486445efde7d505099c15a0a3dba6ef
Needed-By: I10c06d713f6459bb142060fbeedb44601bd2e142