Wait for Vue app to render in Cypress tests
Because our server-rendered markup is so similar to the client-rendered
markup, it was possible for our Cypress tests to partially run against
the SSR without realizing it. For some of the tests, this is actually
fine, but in the edit statement test we really need the links to have
been mounted by Vue.js, with event handlers, before we try to click
them. (Alternatively, perhaps the edit links shouldn’t be in the
server-rendered output at all. But that’s a product decision.)

Bug: T403624
Change-Id: Id299d35cbe7b30b8f1db5c81e3df5e3eb03cdf20