Publish statement edits when 'publish' button is clicked
Persist changes to statements when the wbui2025 edit statement form is
saved.

To support reactivity in changes to statement data, this patch
migrates the statement data from properties to a pinia store and
modifies the components accordingly. In the backend we can use the
StatementLists PHP object to get most of the features we need from
the pinia store (i.e. looking up statements by guid and getting
lists of statements for a given property).

The changes are persisted to Wikibase with calls to `wbeditentity`.
We use `wbformatvalue` to get an updated render of the read-only
data value (and use that to update our `snakValueHtml` store).

In follow-up changes, we plan to refactor StatmentSectionsView to
make the `setupVueTemplateRenderer` function less unwieldy, and to
add some pinia helpers for the jest tests.

Bug: T401405
Change-Id: I6a6a69649eddbf1fad83261cf4cf8acd77c9bb6a