Add qualifier when editing a statement
This commit introduces a form to add a qualifier while editing an
existing statement. A property can be selected, and if its datatype
is `string` it is possible to enter a value. Upon clicking `add`,
a preview of the new qualifier is added underneath any existing
qualifiers. If the statement's edit is saved, the new qualifier is
included and persisted.

Property lookup is extracted out into a standalone component, and is
used in both `WikibaseWbui2025PropertySelector` and the newly added
`WikibaseWbui2025AddQualifier` components.

This is also the first instance of going from one modal to another.
A change was necessary to keep scrolling disabled for the content
hidden under the modals. Rather than teleporting the modal to the
body element, it goes to `mw-teleport-target` instead, which appears
to have been added specifically for modals like these. Now, when
a modal is unmounted, it checks whether any other modals remain before
removing the scroll-blocking body class.

Bug: T402431
Change-Id: Ic00c6b1c6d2a537d87a2adfa646cac3892f686c5