Add message for edit statement dialog header
This removes what I think was the last hard-coded English text in there.
We introduce a new message with a basic singular/plural distinction for
whether one statement or a group of multiple statements is being edited;
as the exact number isn’t shown, I expect even lanugages with e.g.
“dual” or “many” plural forms, won’t use them, but I’m not confident
enough about that to mention it in qqq.json – the translators can figure
it out.

Note that this will show “edit statements” (plural) if the user removes
all statements in the dialog, which may not be the intended behavior.

Bug: T402630
Change-Id: I345b39cd0e6478882846af19d504ee726685c186