fix: Fix spaces in anon edit warnings
Pass `$query` as an array for automatic escaping, and use DB key form
for prettier output (`_` instead of `+`). Either of these could solve
the bug mentioned in the task, but using an array provides an extra
safety net, and using DB key form, as mentioned, results in prettier
URLs. Move the title formatting to AnonymousEditWarningBuilder to ensure
the DB key form.

While we’re at it,

- Instead of using a `MessageLocalizer` unrelated to the special page/
  action calling the service (a possible source of bugs), return a
  `MessageValue`, and format it in the callers with the appropriate
  context.
- Make temp user handling look less like an afterthought: use a ternary
  expression in the service instead of unconditionally assigning the
  no-temp-user value and then overwriting it; and use a data provider
  instead of copy-and-paste in the test.

Deprecate but keep the old signature, because it’s used by
WikibaseLexeme.

Bug: T404499
Change-Id: Ibc3afa6be49758413b498cf53186912aa5be201e