refactor(LanguageAwareRenderer): Use PageReference instead of Title
`TrackingCategories::addTrackingCategory()` accepts any `PageReference`,
not only `Title`s. Per coding conventions [1], `Title` shouldn’t be used
as a type declaration.

The return type of `Parser::getPage()` is documented to be nullable,
so even if it’s unlikely, propagate this nullability; at the end of the
call chain, `TrackingCategories::addTrackingCategory()` also accepts
`null`s (it logs a warning but doesn’t throw exceptions).

[1] https://www.mediawiki.org/wiki/Manual:Coding_conventions/PHP#Don%27t_add_type_declarations_for_%22big%22_legacy_classes

Change-Id: I254407f2383ff314dad0ab9c291eeca9feb3a33d