Serialize data about new statements, remove deleted statements
When the user uses the 'add value' or 'remove' buttons in the
`editStatement` form, those new claims need to be propogated to
the server and the changes reflected in the local store.

Track which values are newly created, and which are removed,
and make the appropriate changes on the backend and locally.

Bug: T401405
Change-Id: Ia399fa935915a0add148b0a3bb22010921637871