tests: Fix ULS stub to work even when ULS is not installed
Follows-up to I99e12bc2cb256, which added a `$.uls = undefined`
assignment so that the property exists (instead of not defined),
after which a sinon.replace (via this.sandbox.stub) is valid.

Between Sinon 1.x and Sinon 18, sinon.stub changed to require a
distinction between define() and replace(). I created the property
with the value undefined so that we can simply call stub/replace instead
of having to call define() sometime and replace() otherwise.

But... Sinon 18+ specifically rejects "undefined" as value,
I suspect to avoid mistakes. Which means we can't use that.

Do my original plan instead, which I actually left in the code
comment: define or replace, depending on whether ULS is installed.

* sinon.define() will create it, and delete it afterward.
* sinon.replace() will replace it, and restore it afterward.
* sinon.stub() is the same as sinon.replace(), but specifically for
  function replacements, where you function override is wrapped
  in a sinon.spy() to e.g. check call count, called with args, etc.

Back in Sinon 1.x, sandbox.stub() basically did all of the above
(if undefined then define it, else replace it; if a function,
wrap with spy.)

Bug: T389450
Bug: T389998
Change-Id: I9d6f28f3fadb8f1614e0ca3ca94b86b3e1efeb70